requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
Authlib==1.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.37.0
cryptography==41.0.5
discord.py==2.3.2
Flask==3.0.0
Flask-Discord==0.1.69
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==2.1.0
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.1.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
